Rethought desired timings
After discussing with YuRow captain, updated the times shown in the table to reflect the times around the weekly sessions instead of just for the next 9 hours and in 24 hours. Fought a bug where using a or t as my for loop variables magically broke the code and also all the related code ABOVE it?? I think my laptop is haunted.